refactor(desktop): extract LSP providers into separate files
