+++ editors/desktop/src/lsp/client.ts
@@ -165,200 +165,13 @@ export class LspClient {
 
     private registerProviders() {
         const languageId = 'lex';
+        if (!this.connection) return;
 
-        // Completion
-        monaco.languages.registerCompletionItemProvider(languageId, {
-            provideCompletionItems: async (model, position, context) => {
-                if (!this.connection) return { suggestions: [] };
-                
-                const params = {
-                    textDocument: { uri: model.uri.toString() },
-                    position: { line: position.lineNumber - 1, character: position.column - 1 },
-                    context: {
-                        triggerKind: context.triggerKind === monaco.languages.CompletionTriggerKind.TriggerCharacter ? 2 : 1,
-                        triggerCharacter: context.triggerCharacter
-                    }
-                };
-
-                try {
-                    const result = await this.connection.sendRequest('textDocument/completion', params) as any;
-                    const items = Array.isArray(result) ? result : result.items;
-                    return {
-                        suggestions: items.map((item: any) => ({
-                            label: item.label,
-                            kind: item.kind ? item.kind - 1 : monaco.languages.CompletionItemKind.Text,
-                            insertText: item.insertText || item.label,
-                            insertTextRules: item.insertTextFormat === 2 ? monaco.languages.CompletionItemInsertTextRule.InsertAsSnippet : undefined,
-                            documentation: item.documentation,
-                            detail: item.detail,
-                            range: item.textEdit ? {
-                                startLineNumber: item.textEdit.range.start.line + 1,
-                                startColumn: item.textEdit.range.start.character + 1,
-                                endLineNumber: item.textEdit.range.end.line + 1,
-                                endColumn: item.textEdit.range.end.character + 1
-                            } : undefined
-                        }))
-                    };
-                } catch (e) {
-                    console.error('[LSP] Completion failed:', e);
-                    return { suggestions: [] };
-                }
-            }
-        });
-
-        // Hover
-        monaco.languages.registerHoverProvider(languageId, {
-            provideHover: async (model, position) => {
-                if (!this.connection) return null;
-                const params = {
-                    textDocument: { uri: model.uri.toString() },
-                    position: { line: position.lineNumber - 1, character: position.column - 1 }
-                };
-                try {
-                    const result = await this.connection.sendRequest('textDocument/hover', params) as any;
-                    if (!result || !result.contents) return null;
-                    return {
-                        contents: Array.isArray(result.contents) 
-                            ? result.contents.map((c: any) => ({ value: typeof c === 'string' ? c : c.value }))
-                            : [{ value: typeof result.contents === 'string' ? result.contents : result.contents.value }]
-                    };
-                } catch (e) {
-                    console.error('[LSP] Hover failed:', e);
-                    return null;
-                }
-            }
-        });
-
-        // Formatting
-        monaco.languages.registerDocumentFormattingEditProvider(languageId, {
-            provideDocumentFormattingEdits: async (model, options) => {
-                if (!this.connection) return [];
-                const params = {
-                    textDocument: { uri: model.uri.toString() },
-                    options: {
-                        tabSize: options.tabSize,
-                        insertSpaces: options.insertSpaces
-                    }
-                };
-                try {
-                    const result = await this.connection.sendRequest('textDocument/formatting', params) as any[];
-                    if (!result) return [];
-                    return result.map(edit => ({
-                        range: {
-                            startLineNumber: edit.range.start.line + 1,
-                            startColumn: edit.range.start.character + 1,
-                            endLineNumber: edit.range.end.line + 1,
-                            endColumn: edit.range.end.character + 1
-                        },
-                        text: edit.newText
-                    }));
-                } catch (e) {
-                    console.error('[LSP] Formatting failed:', e);
-                    return [];
-                }
-            }
-        });
-
-        // Range Formatting
-        monaco.languages.registerDocumentRangeFormattingEditProvider(languageId, {
-            provideDocumentRangeFormattingEdits: async (model, range, options) => {
-                if (!this.connection) return [];
-                const params = {
-                    textDocument: { uri: model.uri.toString() },
-                    range: {
-                        start: { line: range.startLineNumber - 1, character: range.startColumn - 1 },
-                        end: { line: range.endLineNumber - 1, character: range.endColumn - 1 }
-                    },
-                    options: {
-                        tabSize: options.tabSize,
-                        insertSpaces: options.insertSpaces
-                    }
-                };
-                try {
-                    const result = await this.connection.sendRequest('textDocument/rangeFormatting', params) as any[];
-                    if (!result) return [];
-                    return result.map(edit => ({
-                        range: {
-                            startLineNumber: edit.range.start.line + 1,
-                            startColumn: edit.range.start.character + 1,
-                            endLineNumber: edit.range.end.line + 1,
-                            endColumn: edit.range.end.character + 1
-                        },
-                        text: edit.newText
-                    }));
-                } catch (e) {
-                    console.error('[LSP] Range Formatting failed:', e);
-                    return [];
-                }
-            }
-        });
-
-        // Definition
-        monaco.languages.registerDefinitionProvider(languageId, {
-            provideDefinition: async (model, position) => {
-                if (!this.connection) return null;
-                const params = {
-                    textDocument: { uri: model.uri.toString() },
-                    position: { line: position.lineNumber - 1, character: position.column - 1 }
-                };
-                try {
-                    const result = await this.connection.sendRequest('textDocument/definition', params) as any;
-                    if (!result) return null;
-                    const locations = Array.isArray(result) ? result : [result];
-                    return locations.map((loc: any) => ({
-                        uri: monaco.Uri.parse(loc.uri),
-                        range: {
-                            startLineNumber: loc.range.start.line + 1,
-                            startColumn: loc.range.start.character + 1,
-                            endLineNumber: loc.range.end.line + 1,
-                            endColumn: loc.range.end.character + 1
-                        }
-                    }));
-                } catch (e) {
-                    console.error('[LSP] Definition failed:', e);
-                    return null;
-                }
-            }
-        });
-
-        // Semantic Tokens
-        monaco.languages.registerDocumentSemanticTokensProvider(languageId, {
-            getLegend: () => ({
-                tokenTypes: [
-                    'DocumentTitle', 'SessionMarker', 'SessionTitleText', 'DefinitionSubject', 'DefinitionContent',
-                    'ListMarker', 'ListItemText', 'AnnotationLabel', 'AnnotationParameter', 'AnnotationContent',
-                    'InlineStrong', 'InlineEmphasis', 'InlineCode', 'InlineMath', 'Reference', 'ReferenceCitation',
-                    'ReferenceFootnote', 'VerbatimSubject', 'VerbatimLanguage', 'VerbatimAttribute', 'VerbatimContent',
-                    'InlineMarker_strong_start', 'InlineMarker_strong_end', 'InlineMarker_emphasis_start', 'InlineMarker_emphasis_end',
-                    'InlineMarker_code_start', 'InlineMarker_code_end', 'InlineMarker_math_start', 'InlineMarker_math_end',
-                    'InlineMarker_ref_start', 'InlineMarker_ref_end',
-                    // Standard types
-                    'comment', 'string', 'keyword', 'number', 'regexp', 'operator', 'namespace',
-                    'type', 'struct', 'class', 'interface', 'enum', 'typeParameter', 'function',
-                    'method', 'decorator', 'macro', 'variable', 'parameter', 'property', 'label'
-                ],
-                tokenModifiers: ['declaration', 'definition', 'readonly', 'static', 'deprecated', 'abstract', 'async', 'modification', 'documentation', 'defaultLibrary']
-            }),
-            provideDocumentSemanticTokens: async (model) => {
-                console.log('[SemanticTokens] Provider triggered');
-                if (!this.connection) return null;
-                const params = {
-                    textDocument: { uri: model.uri.toString() }
-                };
-                try {
-                    const result = await this.connection.sendRequest('textDocument/semanticTokens/full', params) as any;
-                    if (!result || !result.data) return null;
-                    console.log(`[SemanticTokens] Received tokens: ${result.data.length}`);
-                    return {
-                        data: new Uint32Array(result.data)
-                    };
-                } catch (e) {
-                    console.error('[LSP] Semantic Tokens failed:', e);
-                    return null;
-                }
-            },
-            releaseDocumentSemanticTokens: () => {}
-        });
+        import('./providers/completion').then(m => m.registerCompletionProvider(languageId, this.connection!));
+        import('./providers/hover').then(m => m.registerHoverProvider(languageId, this.connection!));
+        import('./providers/formatting').then(m => m.registerFormattingProvider(languageId, this.connection!));
+        import('./providers/definition').then(m => m.registerDefinitionProvider(languageId, this.connection!));
+        import('./providers/semantic_tokens').then(m => m.registerSemanticTokensProvider(languageId, this.connection!));
     }
 
     public async sendRequest<R, P = any>(method: string, params: P): Promise<R> {
